Add time window validation and index date selection functionality
- Introduced `check_time_windows` function to validate time window configurations, ensuring required parameters are present and correctly formatted.
- Added `select_time_eligible_exposed` and `draw_index_dates_for_control_with_redraw` functions to filter exposed patients based on eligibility criteria and assign index dates to control patients.
- Implemented unit tests for the new functionality, covering various scenarios including edge cases and error handling.
- Removed deprecated functions from `select_cohort_full.py` to streamline the codebase.
